Extract more methods to InventoryIncrementalJobPublicAPI; Add InventoryIncrementalJobAPI
